ADD: best-matched patch searching
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@3410 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/libs/Base/cmtkDataGrid.h
@@ -188,13 +188,13 @@ public:
   }
 
   /// Get offset of a pixel.
-  size_t GetOffsetFromIndex( const int x, const int y, const int z ) const 
+  int GetOffsetFromIndex( const int x, const int y, const int z ) const 
   {
     return x + nextJ * y + nextK * z;
   }
 
   /// Get offset of a pixel.
-  size_t GetOffsetFromIndex( const Self::IndexType& index ) const 
+  int GetOffsetFromIndex( const Self::IndexType& index ) const 
   {
     return index[0] + this->nextJ * index[1] + this->nextK * index[2];
   }
